Create a SCIM 2.0 /Me endpoint and decommission /Users/self.
